requirements.txt
PySide6>=6.6.0
python-xlib>=0.33
Pillow>=10.0.0
pynput>=1.7.6
